requirements.txt
pandas~=2.0.3
PyYAML~=6.0.1
requests~=2.32.3
setuptools~=68.2.0

numpy
pandas
matplotlib
scikit-learn
torch
notebook
pyyaml
